前端初始工程

+++ keep-accounts-springboot/sky-server/src/main/java/com/sky/service/impl/TradeRecordServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 /**
@@ -25,13 +26,19 @@ public class TradeRecordServiceImpl implements TradeRecordService {
     private TradeRecordMapper tradeRecordMapper;
 
     /**
-     * 添加交易记录
+     * 新增交易记录
      *
      * @param tradeRecordDTO
+     * @param userId
      */
-    public void addTradeRecord(TradeRecordDTO tradeRecordDTO) {
+    public void addTradeRecord(TradeRecordDTO tradeRecordDTO, Long userId) {
         TradeRecord tradeRecord = new TradeRecord();
         BeanUtils.copyProperties(tradeRecordDTO, tradeRecord);
+        tradeRecord.setUserId(userId);
+        tradeRecord.setCreateUser(userId);
+        tradeRecord.setUpdateUser(userId);
+        tradeRecord.setUpdateTime(LocalDateTime.now());
+        tradeRecord.setCreateTime(LocalDateTime.now());
         tradeRecordMapper.insert(tradeRecord);
     }
 
@@ -39,12 +46,25 @@ public class TradeRecordServiceImpl implements TradeRecordService {
      * 删除交易记录
      *
      * @param id
+     * @param userId
      */
-    public void deleteTradeRecord(Long id) {
-        tradeRecordMapper.delete(id);
+    public void deleteTradeRecord(Long id, Long userId) {
+        // 根据主键id查询交易记录的user_id
+        Long userIdInDatabase = tradeRecordMapper.getUserIdById(id);
+        if (userIdInDatabase.equals(userId)) {
+            tradeRecordMapper.delete(id);
+        } else {
+            throw new RuntimeException("删除失败，权限不足");
+        }
     }
 
-    public void batchDeleteTradeRecord(List<Long> ids) {
+    /**
+     * 批量删除交易记录
+     *
+     * @param ids
+     * @param userId
+     */
+    public void batchDeleteTradeRecord(List<Long> ids, Long userId) {
         log.info("批量删除交易记录：{}", ids);
         tradeRecordMapper.batchDelete(ids);
     }
